update to Flyway example

+++ intro/spring/flyway/src/main/resources/db/migration/V1.0__createDB.sql
@@ -1,5 +1,5 @@
 create table First (
-  id numeric,
+  id BIGINT,
   data  varchar(50),
   constraint pk_First primary key (id)
 );
