+++ intro/spring/flyway/src/main/resources/db/migration/V1.1__addSecondAndChangeFirst.sql
@@ -1,5 +1,5 @@
 create table Second (
-  id numeric,
+  id BIGINT,
   data  varchar(50),
   constraint pk_Second primary key (id)
 );
